+++ pyproject.toml
@@ -8,7 +8,7 @@ readme = "README.md"
 repository = "https://github.com/cognitedata/inso-extpipes-cli"
 
 packages = [
-    { include="src" },
+    { include="extpipes", from="src" },
 ]
 
 [tool.poetry.group.dev.dependencies]
@@ -49,7 +49,7 @@ pytest = "^7.2.1"
 # pytest-cov = "^3.0.0"
 
 [tool.poetry.scripts]
-extpipes-cli = "src.__main__:main"
+extpipes-cli = "extpipes.__main__:main"
 
 [build-system]
 requires = ["poetry>=0.12"]
@@ -60,5 +60,5 @@ version_toml = [
 "pyproject.toml:tool.poetry.version"
 ]
 version_variable = [
-    "src/__init__.py:__version__",
+    "src/extpipes/__init__.py:__version__",
 ]
