+++ src/extpipes/commands/base.py
